refactor: user, common 패키지 리팩터링
