+++ src/main/java/com/example/gistcompetitioncnserver/answer/presentation/AnswerController.java
@@ -1,11 +1,11 @@
 package com.example.gistcompetitioncnserver.answer.presentation;
 
-import com.example.gistcompetitioncnserver.answer.dto.AnswerRequest;
 import com.example.gistcompetitioncnserver.answer.application.AnswerService;
 import com.example.gistcompetitioncnserver.answer.domain.Answer;
+import com.example.gistcompetitioncnserver.answer.dto.AnswerRequest;
 import com.example.gistcompetitioncnserver.config.annotation.LoginUser;
 import com.example.gistcompetitioncnserver.config.annotation.ManagerPermissionRequired;
-import com.example.gistcompetitioncnserver.user.SimpleUser;
+import com.example.gistcompetitioncnserver.user.domain.SimpleUser;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
