+++ src/main/java/com/example/gistcompetitioncnserver/common/password/BcryptEncoder.java
@@ -1,10 +1,10 @@
-package com.example.gistcompetitioncnserver.user;
+package com.example.gistcompetitioncnserver.common.password;
 
 import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.stereotype.Component;
 
 @Component
-public class BcryptEncoder implements Encryptor {
+public class BcryptEncoder implements Encoder {
     @Override
     public String hashPassword(String raw) {
         return BCrypt.hashpw(raw, BCrypt.gensalt());
