+++ src/main/java/com/example/gistcompetitioncnserver/post/application/PostService.java
@@ -6,8 +6,8 @@ import com.example.gistcompetitioncnserver.exception.user.NoSuchUserException;
 import com.example.gistcompetitioncnserver.post.domain.Post;
 import com.example.gistcompetitioncnserver.post.domain.PostRepository;
 import com.example.gistcompetitioncnserver.post.dto.PostRequest;
-import com.example.gistcompetitioncnserver.user.User;
-import com.example.gistcompetitioncnserver.user.UserRepository;
+import com.example.gistcompetitioncnserver.user.domain.User;
+import com.example.gistcompetitioncnserver.user.domain.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Sort;
