--- src/main/java/com/example/gistcompetitioncnserver/user/LoginService.java
@@ -1,9 +1,0 @@
-package com.example.gistcompetitioncnserver.user;
-
-public interface LoginService {
-    void login(SignInRequest request);
-
-    void logout();
-
-    SimpleUser getLoginUser();
-}
