+++ src/main/java/com/example/gistcompetitioncnserver/user/application/SessionLoginService.java
@@ -1,7 +1,12 @@
-package com.example.gistcompetitioncnserver.user;
+package com.example.gistcompetitioncnserver.user.application;
 
+import com.example.gistcompetitioncnserver.common.password.Encoder;
 import com.example.gistcompetitioncnserver.exception.user.NoSuchUserException;
 import com.example.gistcompetitioncnserver.exception.user.NotMatchedPasswordException;
+import com.example.gistcompetitioncnserver.user.domain.SimpleUser;
+import com.example.gistcompetitioncnserver.user.domain.User;
+import com.example.gistcompetitioncnserver.user.domain.UserRepository;
+import com.example.gistcompetitioncnserver.user.dto.request.SignInRequest;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,14 +19,14 @@ public class SessionLoginService implements LoginService {
     public static final String SESSION_KEY = "user";
     private final HttpSession httpSession;
     private final UserRepository userRepository;
-    private final Encryptor encryptor;
+    private final Encoder encoder;
 
     @Override
     @Transactional
     public void login(SignInRequest request) {
         User user = userRepository.findByUsername(request.getUsername())
                 .orElseThrow(NoSuchUserException::new);
-        if (!encryptor.isMatch(request.getPassword(), user.getPassword())) {
+        if (!encoder.isMatch(request.getPassword(), user.getPassword())) {
             throw new NotMatchedPasswordException();
         }
         httpSession.setAttribute(SESSION_KEY, new SimpleUser(user));
