+++ src/main/java/com/example/gistcompetitioncnserver/user/application/SignUpValidator.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.user;
+package com.example.gistcompetitioncnserver.user.application;
 
 public interface SignUpValidator {
     void checkIsVerified(String username, String verificationCode);
