+++ src/test/java/com/example/gistcompetitioncnserver/answer/AnswerServiceTest.java
@@ -10,9 +10,9 @@ import com.example.gistcompetitioncnserver.exception.post.NoSuchPostException;
 import com.example.gistcompetitioncnserver.exception.post.UnAnsweredPostException;
 import com.example.gistcompetitioncnserver.post.domain.Post;
 import com.example.gistcompetitioncnserver.post.domain.PostRepository;
-import com.example.gistcompetitioncnserver.user.User;
-import com.example.gistcompetitioncnserver.user.UserRepository;
-import com.example.gistcompetitioncnserver.user.UserRole;
+import com.example.gistcompetitioncnserver.user.domain.User;
+import com.example.gistcompetitioncnserver.user.domain.UserRepository;
+import com.example.gistcompetitioncnserver.user.domain.UserRole;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
