+++ src/test/java/com/example/gistcompetitioncnserver/post/PostServiceTest.java
@@ -8,9 +8,9 @@ import com.example.gistcompetitioncnserver.post.domain.AgreementRepository;
 import com.example.gistcompetitioncnserver.post.domain.Post;
 import com.example.gistcompetitioncnserver.post.domain.PostRepository;
 import com.example.gistcompetitioncnserver.post.dto.PostRequest;
-import com.example.gistcompetitioncnserver.user.User;
-import com.example.gistcompetitioncnserver.user.UserRepository;
-import com.example.gistcompetitioncnserver.user.UserRole;
+import com.example.gistcompetitioncnserver.user.domain.User;
+import com.example.gistcompetitioncnserver.user.domain.UserRepository;
+import com.example.gistcompetitioncnserver.user.domain.UserRole;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
